chore(postgresql): rename app folder and add icon
