+++ packages/backend/src/apps/postgresql/auth/is-still-verified.ts
@@ -1,0 +1,10 @@
+import { IGlobalVariable } from '@automatisch/types';
+import verifyCredentials from './verify-credentials';
+
+const isStillVerified = async ($: IGlobalVariable) => {
+  await verifyCredentials($);
+
+  return true;
+};
+
+export default isStillVerified;
